Configure macOS to prefer English
The first number in this file is supposed to be the character encoding.
Google tells me that 0x00 probably means MacRoman, but I can't confirm
that; I guess it rather means "default", which seems to be UTF-8 in
newer macOS releases.

The second number is the system language, where 0x03 is apparently
German and 0x00 English. However, I somewhat doubt that this file is
authoritative regarding that setting; I rather expect it to just mirror
whatever is set in some plist. But let's see.