Remove fish config
While I liked fish’s syntax highlighting, I’m not happy with the fact
that it’s not sh compatible. And syntax highlighting alone is not worth
installing a custom shell on all of my machines. Therefore, fish has to
go. I’m sorry, fish. So long, and thanks for all the … you know.