Automatic WSL ssh-agent forwarding to gpg-agent
Had to switch to a tmux background session because all of the
setsid/nohup/&/disown attempts did not really work.

wsl-ssh-pageant is really good at quitting without removing its socket,
or simply not quitting at all, so I had to add a check for whether or
not the socket works.

Also, apparently socat can also die without removing the socket, so I've
added the same check there as well.